Fixed setPixel to wrong index in frame buffer. The index to the buffer array is integer and not float. Because of it there was bad access to the buffer and wrong pixels were set.

+++ source/cpp/helpers/DrawingLibrary_DEPTH.h
@@ -146,7 +146,7 @@ class DrawingLibraryDepth : public DrawingLibraryInterface
 
 		inline void setPixel(float x, float y, float z, Chunk &context)
 		{
-			uint_fast32_t index = uint_fast32_t(x + context.w * y);
+			uint_fast32_t index = uint_fast32_t((int)x + context.w * (int)y);
 			if (x >= 0 && x < context.w && y >= 0 && y < context.h && context.depth[index] > z)
 			{
 				context.lastSetPixelIndex	= index;
